ISPN-3458 Move the HotRod target migrator to the remotecachestore - Also drop the upgrade-tools module as its become useless now
