+++ cachestore/remote/src/main/resources/META-INF/services/org.infinispan.upgrade.TargetMigrator
@@ -1,0 +1,1 @@
+org.infinispan.loaders.remote.upgrade.HotRodTargetMigrator
