+++ cli/cli-server/pom.xml
@@ -83,11 +83,6 @@
          <scope>test</scope>
       </dependency>
       <dependency>
-         <groupId>${project.groupId}</groupId>
-         <artifactId>infinispan-upgrade-tools</artifactId>
-         <scope>test</scope>
-      </dependency>
-      <dependency>
          <groupId>spy</groupId>
          <artifactId>memcached</artifactId>
          <scope>test</scope>
