+++ pom.xml
@@ -42,7 +42,6 @@
       <module>server/rest</module>
       <module>client/hotrod-client</module>
       <module>rhq-plugin</module>
-      <module>upgrade-tools</module>
       <module>spring</module>
       <module>cli/cli-server</module>
       <module>cli/cli-client</module>
