--- upgrade-tools/pom.xml
@@ -1,74 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-   <modelVersion>4.0.0</modelVersion>
-   <parent>
-      <groupId>org.infinispan</groupId>
-      <artifactId>infinispan-parent</artifactId>
-      <version>6.0.0-SNAPSHOT</version>
-      <relativePath>../parent/pom.xml</relativePath>
-   </parent>
-   
-   <artifactId>infinispan-upgrade-tools</artifactId>
-   <name>Infinispan Rolling Upgrade Tooling</name>
-
-   <dependencies>
-      <dependency>
-         <groupId>org.infinispan</groupId>
-         <artifactId>infinispan-client-hotrod</artifactId>
-      </dependency>
-      
-      <dependency>
-         <groupId>org.infinispan</groupId>
-         <artifactId>infinispan-cachestore-remote</artifactId>
-      </dependency>
-      
-      <dependency>
-         <groupId>${project.groupId}</groupId>
-         <artifactId>infinispan-core</artifactId>
-         <type>test-jar</type>
-         <scope>test</scope>
-      </dependency>
-      
-      <dependency>
-         <groupId>${project.groupId}</groupId>
-         <artifactId>infinispan-client-hotrod</artifactId>
-         <type>test-jar</type>
-         <scope>test</scope>
-      </dependency>
-      
-      <dependency>
-         <groupId>${project.groupId}</groupId>
-         <artifactId>infinispan-server-hotrod</artifactId>
-         <scope>test</scope>
-      </dependency>
-      
-      <dependency>
-         <groupId>${project.groupId}</groupId>
-         <artifactId>infinispan-server-hotrod</artifactId>
-         <type>test-jar</type>
-         <scope>test</scope>
-      </dependency>
-   </dependencies>
-   
-   <build>
-      <plugins>
-         <plugin>
-            <groupId>org.antlr</groupId>
-            <artifactId>antlr3-maven-plugin</artifactId>
-         </plugin>
-         <plugin>
-            <groupId>org.apache.felix</groupId>
-            <artifactId>maven-bundle-plugin</artifactId>
-            <configuration>
-               <instructions>
-                  <Export-Package>
-                     ${project.groupId}.upgrade.*;version=${project.version};-split-package:=error
-                  </Export-Package>
-               </instructions>
-            </configuration>
-         </plugin>
-      </plugins>
-   </build>
-</project>
