--- upgrade-tools/src/main/resources/META-INF/services/org.infinispan.upgrade.TargetMigrator
@@ -1,1 +1,0 @@
-org.infinispan.upgrade.hotrod.HotRodTargetMigrator
